requirements.txt
absl-py==2.3.1
altair==5.5.0
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.60.2
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.78.0
h5py==3.14.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.10.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
narwhals==2.18.0
numpy==2.0.2
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.5
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
referencing==0.36.2
requests==2.32.5
rich==14.3.3
rpds-py==0.27.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
smmap==5.0.3
streamlit==1.50.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.6
wrapt==2.1.2
zipp==3.23.0
